Move functionality for parsing files to ave.parsing
