+++ ave/__init__.py
@@ -1,7 +1,7 @@
 """The AVE game engine."""
 
 from .ave import AVE  # noqa: F401
-from .game_loader import load_game_from_library  # noqa: F401
-from .game_loader import load_game_from_file  # noqa: F401
+from .parsing.game_loader import load_game_from_library  # noqa: F401
+from .parsing.game_loader import load_game_from_file  # noqa: F401
 from .game import Character, Game  # noqa: F401
 from .config import version as __version__  # noqa: F401
