+++ ave/parsing/file_parsing.py
@@ -1,14 +1,14 @@
 """Parsing a .ave file."""
 
 import re
-from .game import Room
-from .components import (TextWithRequirements, OptionWithRequirements,
-                         NameWithRequirements)
+from ..game import Room
+from ..components import (TextWithRequirements, OptionWithRequirements,
+                          NameWithRequirements)
 from .string_functions import escape, unescape, clean, between
-from .components.items import Item, NumberItem
-from .components import requirements as rq
-from .components import item_giver as ig
-from .components import numbers as no
+from ..components.items import Item, NumberItem
+from ..components import requirements as rq
+from ..components import item_giver as ig
+from ..components import numbers as no
 
 
 def _parse_rq(condition):
